move TraitDid and Krate to utils/trait_did.rs
